Refactor license actions rendering logic and move it to a separate component folder.
